update difficulty types

+++ app/models/difficulty.py
@@ -2,6 +2,6 @@ from pydantic import BaseModel
 from typing import Tuple
 
 class Difficulty(BaseModel):
-    input_state: Tuple[float, int, int]
+    input_state: Tuple[float, float, int]
     action_taken: int
     adjusted_difficulty: int
